move contents of `mars-rover-health-types` into `mars_types::health`
