--- packages/health-types/examples/schema.rs
@@ -1,10 +1,0 @@
-use cosmwasm_schema::write_api;
-use mars_rover_health_types::{ExecuteMsg, InstantiateMsg, QueryMsg};
-
-fn main() {
-    write_api! {
-        instantiate: InstantiateMsg,
-        query: QueryMsg,
-        execute: ExecuteMsg,
-    }
-}
