+++ packages/health/src/health.rs
@@ -2,6 +2,7 @@ use std::{collections::HashMap, fmt};
 
 use cosmwasm_std::{Addr, Coin, Decimal, Fraction, QuerierWrapper, StdResult, Uint128};
 use mars_red_bank_types::params::AssetParams;
+use mars_types::health::HealthValuesResponse;
 
 use crate::{error::HealthError, query::MarsQuerier};
 
@@ -43,6 +44,19 @@ impl fmt::Display for Health {
             self.max_ltv_health_factor.map_or("n/a".to_string(), |x| x.to_string()),
             self.liquidation_health_factor.map_or("n/a".to_string(), |x| x.to_string())
         )
+    }
+}
+
+impl From<HealthValuesResponse> for RedBankHealth {
+    fn from(h: HealthValuesResponse) -> Self {
+        Self {
+            total_debt_value: h.total_debt_value,
+            total_collateral_value: h.total_collateral_value,
+            max_ltv_adjusted_collateral: h.max_ltv_adjusted_collateral,
+            liquidation_threshold_adjusted_collateral: h.liquidation_threshold_adjusted_collateral,
+            max_ltv_health_factor: h.max_ltv_health_factor,
+            liquidation_health_factor: h.liquidation_health_factor,
+        }
     }
 }
