+++ packages/types/src/health/error.rs
